feat: global config file to store persistent user info
The default location for this file is in the user's home directory with
the file name 'cdk-today.json'.
The first entry in the global config file is 'user' which holds the
Github username.

The system detects when the config file is missing and initiates a setup
procedure which prompts the user to provide this information. This also
detects and fails when the file is poorly configured.