chore: code simplification
Merge PullRequest and PullRequests into Issue and Issues
correspondingly.

Merge TableView into Issues.

The motivation of is to,
a. reduce the amount of code duplication across pullrequests.ts and
issues.ts, and
b. reduce the amount of clutter with the decoupling of tableview and
issues. This decoupling felt like premature optimization whose future
value was not clear, with the added overhead of keeping the
implementation abstract.